birthday schema, leave approval, and change in employee module

+++ src/modules/auth/repository.ts
@@ -1,39 +1,28 @@
+// src/modules/auth/repository.ts
 import { prisma } from "../../config/prisma.js";
 
 export class AuthRepository {
-  // ---------Company---------
-  async findCompanyById(companyId: string) {
+  findCompanyById(companyId: string) {
     return prisma.company.findUnique({
       where: { id: companyId },
-      select: {
-        id: true,
-        isActive: true,
-      },
+      select: { id: true, isActive: true },
     });
   }
 
-  // ---------- USERS ----------
-
   findUserByEmail(email: string) {
-    return prisma.user.findUnique({
-      where: { email },
-    });
+    return prisma.user.findUnique({ where: { email } });
   }
 
   findUserById(userId: string) {
-    return prisma.user.findUnique({
-      where: { id: userId },
-    });
+    return prisma.user.findUnique({ where: { id: userId } });
   }
-
-  // ---------- REFRESH TOKENS ----------
 
   createRefreshToken(params: {
     userId: string;
     token: string;
     expiresAt: Date;
-    userAgent?: string
-    ipAddress?: string
+    userAgent?: string;
+    ipAddress?: string;
   }) {
     return prisma.refreshToken.create({
       data: {
@@ -49,28 +38,15 @@ export class AuthRepository {
   findRefreshToken(token: string) {
     return prisma.refreshToken.findUnique({
       where: { token },
-      include: {
-        user: true,
-      },
+      include: { user: true },
     });
   }
 
   deleteRefreshToken(token: string) {
-    return prisma.refreshToken.delete({
-      where: { token },
-    });
+    return prisma.refreshToken.delete({ where: { token } });
   }
 
-  deleteAllRefreshTokensForUser(userId: string) {
-    return prisma.refreshToken.deleteMany({
-      where: { userId },
-    });
+  deleteAllRefreshTokensByUser(userId: string) {
+    return prisma.refreshToken.deleteMany({ where: { userId } });
   }
-
-  async deleteAllRefreshTokensByUser(userId: string) {
-  return prisma.refreshToken.deleteMany({
-    where: { userId },
-  })
-}
-
 }
